+++ apps/vscode/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Implement support for Go to Definition within code cells
 - Update to Mermaid JS v9.3.0 (prevent preview errors for YAML front matter)
+- Only show Quarto Assist panel when the extension is activated
 
 ## 1.62.0 (Release on 8 February 2023)